+++ src/timer/delay.rs
@@ -1,124 +1,93 @@
 //! Delays
 
-use super::{Error, FTimer, General, Instance, Timer};
-use crate::pac;
+use super::{FTimer, Instance, Timer};
 use core::ops::{Deref, DerefMut};
 use cortex_m::peripheral::SYST;
-use fugit::TimerDurationU32;
+use fugit::{MicrosDurationU32, TimerDurationU32};
 
 /// Timer as a delay provider (SysTick by default)
-pub struct Delay<TIM = SYST>(Timer<TIM>);
+pub struct SysDelay(Timer<SYST>);
 
-impl<T> Deref for Delay<T> {
-    type Target = Timer<T>;
+impl Deref for SysDelay {
+    type Target = Timer<SYST>;
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl<T> DerefMut for Delay<T> {
+impl DerefMut for SysDelay {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
 }
 
-impl Delay<SYST> {
+impl SysDelay {
     /// Releases the timer resource
     pub fn release(self) -> Timer<SYST> {
         self.0
     }
 }
 
-impl<TIM: Instance> Delay<TIM> {
-    /// Releases the timer resource
-    pub fn release(mut self) -> Timer<TIM> {
-        self.tim.cr1_reset();
-        self.0
+impl Timer<SYST> {
+    pub fn delay(self) -> SysDelay {
+        SysDelay(self)
     }
 }
 
-impl<TIM> Timer<TIM> {
-    pub fn delay(self) -> Delay<TIM> {
-        Delay(self)
+impl SysDelay {
+    pub fn delay(&mut self, us: MicrosDurationU32) {
+        // The SysTick Reload Value register supports values between 1 and 0x00FFFFFF.
+        const MAX_RVR: u32 = 0x00FF_FFFF;
+
+        let mut total_rvr = us.ticks() * (self.clk.raw() / 1_000_000);
+
+        while total_rvr != 0 {
+            let current_rvr = total_rvr.min(MAX_RVR);
+
+            self.tim.set_reload(current_rvr);
+            self.tim.clear_current();
+            self.tim.enable_counter();
+
+            // Update the tracking variable while we are waiting...
+            total_rvr -= current_rvr;
+
+            while !self.tim.has_wrapped() {}
+
+            self.tim.disable_counter();
+        }
     }
-}
-
-mod sealed {
-    pub trait Wait {
-        fn wait(&mut self, prescaler: u16, auto_reload_register: u32);
-    }
-}
-pub(super) use sealed::Wait;
-
-macro_rules! hal {
-    ($($TIM:ty,)+) => {
-        $(
-            impl Wait for Delay<$TIM> {
-                fn wait(&mut self, prescaler: u16, auto_reload_register: u32) {
-                    // Write Prescaler (PSC)
-                    self.tim.set_prescaler(prescaler - 1);
-
-                    // Write Auto-Reload Register (ARR)
-                    // Note: Make it impossible to set the ARR value to 0, since this
-                    // would cause an infinite loop.
-                    self.tim.set_auto_reload(auto_reload_register - 1).unwrap();
-
-                    // Trigger update event (UEV) in the event generation register (EGR)
-                    // in order to immediately apply the config
-                    self.tim.trigger_update();
-
-                    // Configure the counter in one-pulse mode (counter stops counting at
-                    // the next updateevent, clearing the CEN bit) and enable the counter.
-                    self.tim.start_one_pulse();
-
-                    // Wait for CEN bit to clear
-                    while self.tim.is_counter_enabled() { /* wait */ }
-                }
-            }
-        )+
-    }
-}
-
-hal! {
-    pac::TIM5,
-}
-
-#[cfg(feature = "tim2")]
-hal! {
-    pac::TIM2,
 }
 
 /// Periodic non-blocking timer that imlements [embedded_hal::blocking::delay] traits
-pub struct FDelay<TIM, const FREQ: u32>(pub(super) FTimer<TIM, FREQ>);
+pub struct Delay<TIM, const FREQ: u32>(pub(super) FTimer<TIM, FREQ>);
 
-impl<T, const FREQ: u32> Deref for FDelay<T, FREQ> {
+impl<T, const FREQ: u32> Deref for Delay<T, FREQ> {
     type Target = FTimer<T, FREQ>;
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl<T, const FREQ: u32> DerefMut for FDelay<T, FREQ> {
+impl<T, const FREQ: u32> DerefMut for Delay<T, FREQ> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
 }
 
-/// `FDelay` with precision of 1 μs (1 MHz sampling)
-pub type DelayUs<TIM> = FDelay<TIM, 1_000_000>;
+/// `Delay` with precision of 1 μs (1 MHz sampling)
+pub type DelayUs<TIM> = Delay<TIM, 1_000_000>;
 
-/// `FDelay` with precision of 1 ms (1 kHz sampling)
+/// `Delay` with precision of 1 ms (1 kHz sampling)
 ///
 /// NOTE: don't use this if your system frequency more than 65 MHz
-pub type DelayMs<TIM> = FDelay<TIM, 1_000>;
+pub type DelayMs<TIM> = Delay<TIM, 1_000>;
 
-impl<TIM: Instance, const FREQ: u32> FDelay<TIM, FREQ> {
+impl<TIM: Instance, const FREQ: u32> Delay<TIM, FREQ> {
     /// Sleep for given time
     pub fn delay(&mut self, time: TimerDurationU32<FREQ>) {
-        let mut ticks = time.ticks() - 1;
-        while ticks > 0 {
+        let mut ticks = time.ticks().min(1) - 1;
+        while ticks != 0 {
             let reload = ticks.min(TIM::max_auto_reload());
-            ticks -= reload;
 
             // Write Auto-Reload Register (ARR)
             unsafe {
@@ -133,6 +102,8 @@ impl<TIM: Instance, const FREQ: u32> FDelay<TIM, FREQ> {
             // the next updateevent, clearing the CEN bit) and enable the counter.
             self.tim.start_one_pulse();
 
+            // Update the tracking variable while we are waiting...
+            ticks -= reload;
             // Wait for CEN bit to clear
             while self.tim.is_counter_enabled() { /* wait */ }
         }
@@ -150,10 +121,19 @@ impl<TIM: Instance, const FREQ: u32> FDelay<TIM, FREQ> {
     }
 }
 
-impl<TIM: Instance, const FREQ: u32> fugit_timer::Delay<FREQ> for FDelay<TIM, FREQ> {
-    type Error = Error;
+impl<TIM: Instance, const FREQ: u32> fugit_timer::Delay<FREQ> for Delay<TIM, FREQ> {
+    type Error = core::convert::Infallible;
 
     fn delay(&mut self, duration: TimerDurationU32<FREQ>) -> Result<(), Self::Error> {
+        self.delay(duration);
+        Ok(())
+    }
+}
+
+impl fugit_timer::Delay<1_000_000> for SysDelay {
+    type Error = core::convert::Infallible;
+
+    fn delay(&mut self, duration: MicrosDurationU32) -> Result<(), Self::Error> {
         self.delay(duration);
         Ok(())
     }